Moved tests into a separate module.
